+++ app/src/main/java/com/example/rickandmorty/data/repository/ItemRepositoryImpl.kt
@@ -16,16 +16,16 @@ class ItemRepositoryImpl(private val exampleDao: ExampleDao) : ItemRepository {
             .subscribeOn(Schedulers.io())
     }
 
-    override fun insertExample(example: ExampleEntity): Observable<Resource<Long>> {
+    override fun insertExample(example: ExampleEntity): Observable<Resource<Int>> {
         return exampleDao.insertExample(example)
             .toObservable()
-            .map<Resource<Long>> { Resource.Data(it) }
+            .map<Resource<Int>> { Resource.Data(it) }
             .onErrorReturn { Resource.Error(it) }
             .startWith(Resource.Loading)
             .subscribeOn(Schedulers.io())
     }
 
-    override fun deleteExample(id: Long): Observable<Resource<Unit>> {
+    override fun deleteExample(id: Int): Observable<Resource<Unit>> {
         return exampleDao.deleteExample(id)
             .andThen(Observable.just(Resource.Success))
             .onErrorReturn { Resource.Error(it) }
